pyproject.toml
[tool.poetry]
name = "notification_module"
version = "0.1.0"
description = ""
authors = ["Alemkhan <alemkhan.yergaliev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.36"
asyncpg = "^0.25.0"
loguru = "^0.6.0"
redis = "^4.3.1"
python-dotenv = "^0.20.0"
orjson = "^3.6.8"
alembic = "^1.7.7"
PyJWT = "^2.4.0"


[tool.poetry.dev-dependencies]
mypy = "^0.950"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
sqlalchemy2-stubs = "^0.0.2-alpha.22"
types-redis = "^4.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
